+++ shared/src/index.ts
@@ -2,5 +2,5 @@ export * from "./types.js";
 export * from "./constants.js";
 export * from "./messages.js";
 export * from "./data/creatures.js";
-export * from "./data/recipes.js";
 export * from "./data/shapes.js";
+export * from "./data/progression.js";